pyproject.toml
[project]
name = "openprobe"
version = "1.0.0"
description = "OpenProbe"

dependencies = [
    "openai==1.82.1",
    "langchain==0.3.24", 
    "langchain-core==0.3.56",
    "langgraph==0.2.76",
    "langchain-google-genai==2.1.3",
    "langchain-community==0.3.23",
    "langchain-openai==0.3.15",
    "langchain_experimental==0.3.5rc1",
    "google-ai-generativelanguage==0.6.18",
    "python-dotenv==1.1.0", 
    "crawl4ai @ git+https://github.com/salzubi401/crawl4ai.git@main",
    "wikipedia-api==0.8.1",
    "boto3==1.38.27",
    "loguru==0.7.3",
    "datasets==3.6.0",
    "mistralai==1.8.1",
    "fasttext-wheel==0.9.2",
    "numpy==1.26.4",
    "cchardet==2.1.7"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0"}

[tool.hatch.build.targets.wheel]
packages = ["src/deepsearch"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pdm]
distribution = true

[tool.hatch.metadata]
allow-direct-references = true
